chore: create session listener
