+++ griffindor-backend/src/main/java/com/devathon/griffindor_backend/controllers/GameSessionController.java
@@ -2,22 +2,16 @@ package com.devathon.griffindor_backend.controllers;
 
 import com.devathon.griffindor_backend.config.WebSocketRoutes;
 import com.devathon.griffindor_backend.dtos.TokenIdResponseDto;
-import com.devathon.griffindor_backend.enums.PlayerSessionState;
 import com.devathon.griffindor_backend.services.ErrorService;
 import com.devathon.griffindor_backend.services.PlayerService;
 import com.devathon.griffindor_backend.utils.Jwt;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.event.EventListener;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
 import org.springframework.messaging.simp.annotation.SendToUser;
-import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.socket.messaging.SessionConnectedEvent;
-import org.springframework.web.socket.messaging.SessionDisconnectEvent;
-import org.springframework.web.socket.messaging.SessionSubscribeEvent;
 
 @RequiredArgsConstructor
 @Controller
@@ -28,33 +22,7 @@ public class GameSessionController {
     private final Jwt jwt;
     private final ErrorService errorService;
 
-    @EventListener
-    public void handleWebSocketConnectListener(SessionConnectedEvent event) {
-        String sessionId = getSessionIdFromEvent(event);
-        playerService.addPlayer(sessionId, null, null, PlayerSessionState.CONNECT, null);
-    }
-
-    @EventListener
-    public void handleWebSocketDisconnectListener(SessionDisconnectEvent event) {
-        String sessionId = getSessionIdFromEvent(event);
-        playerService.updatePlayerSessionState(sessionId, PlayerSessionState.DISCONNECT);
-    }
-
-    @EventListener
-    public void handleSubscribeEvent(SessionSubscribeEvent event) {
-        StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
-
-        String sessionId = headerAccessor.getSessionId();
-        String destination = headerAccessor.getDestination();
-
-        if (WebSocketRoutes.TOPIC_NUM_PLAYERS.equals(destination)) {
-            playerService.enqueuePlayersConnectedEvent(sessionId);
-
-        }
-        if (WebSocketRoutes.USER_QUEUE_LIST_PLAYERS.equals(destination)) {
-            playerService.enqueuePlayersListEvent(sessionId);
-        }
-    }
+   
 
     @MessageMapping(WebSocketRoutes.TOKEN_ID)
     @SendToUser(WebSocketRoutes.QUEUE_TOKEN_ID)
@@ -93,11 +61,5 @@ public class GameSessionController {
         return new TokenIdResponseDto(newToken);
     }
 
-    private String getSessionIdFromEvent(SessionConnectedEvent event) {
-        return StompHeaderAccessor.wrap(event.getMessage()).getSessionId();
-    }
 
-    private String getSessionIdFromEvent(SessionDisconnectEvent event) {
-        return StompHeaderAccessor.wrap(event.getMessage()).getSessionId();
-    }
 }
